Rename `regexp/no-useless-non-greedy` to `regexp/no-useless-lazy`
