+++ lib/utils/rules.ts
@@ -30,6 +30,7 @@ import noUselessDollarReplacements from "../rules/no-useless-dollar-replacements
 import noUselessEscape from "../rules/no-useless-escape"
 import noUselessExactlyQuantifier from "../rules/no-useless-exactly-quantifier"
 import noUselessFlag from "../rules/no-useless-flag"
+import noUselessLazy from "../rules/no-useless-lazy"
 import noUselessNonCapturingGroup from "../rules/no-useless-non-capturing-group"
 import noUselessNonGreedy from "../rules/no-useless-non-greedy"
 import noUselessRange from "../rules/no-useless-range"
@@ -86,6 +87,7 @@ export const rules = [
     noUselessEscape,
     noUselessExactlyQuantifier,
     noUselessFlag,
+    noUselessLazy,
     noUselessNonCapturingGroup,
     noUselessNonGreedy,
     noUselessRange,
